# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SkillHub API",
    "version": "1.0.0",
    "description": "API de gestion pour la plateforme SkillHub. Épreuve EC04."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/workshops": {
      "get": {
        "summary": "Liste des ateliers (avec pagination et filtres)",
        "tags": [
          "Workshops"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste récupérée"
          }
        }
      }
    },
    "/api/workshops/{id}/participants": {
      "get": {
        "summary": "Liste des inscrits à un atelier (Propriétaire uniquement)",
        "tags": [
          "Workshops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des participants"
          },
          "403": {
            "description": "Non autorisé"
          }
        }
      }
    },
    "/api/registrations/{id}/cancel": {
      "patch": {
        "summary": "Annuler une inscription (Apprenant propriétaire uniquement)",
        "tags": [
          "Registrations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inscription annulée"
          },
          "403": {
            "description": "Non autorisé"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Créer un compte utilisateur",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "first_name",
                  "last_name",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "formateur@skillhub.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!",
                    "description": "Minimum 8 caractères, 1 majuscule, 1 chiffre"
                  },
                  "first_name": {
                    "type": "string",
                    "example": "Jean"
                  },
                  "last_name": {
                    "type": "string",
                    "example": "Dupont"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "learner",
                      "trainer"
                    ],
                    "example": "trainer"
                  },
                  "bio": {
                    "type": "string",
                    "example": "Développeur fullstack depuis 10 ans"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé"
          },
          "400": {
            "description": "Erreur de validation"
          },
          "409": {
            "description": "Email déjà utilisé"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Se connecter à l'API",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "chloe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie, retourne le token JWT"
          },
          "400": {
            "description": "Requête invalide"
          },
          "401": {
            "description": "Identifiants incorrects"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentification des utilisateurs"
    }
  ]
}